moved command registration from ready behaviour to handle-commands behaviour
